fix gear value bug

+++ data_parser.h
@@ -13,7 +13,7 @@ class DataParser
     uint16_t getExpectedSteeringPosition();
     uint16_t getExpectedBrakePosition();
     uint16_t getExpectedAcceleratorPosition();
-    uint8_t  getExpectedGearPosition();
+    uint16_t  getExpectedGearPosition();
     uint8_t  getExpectedAutonomyStatus();
     uint8_t  getExpectedIgnitionStatus();
 
@@ -60,7 +60,6 @@ void DataParser::parseExternalData(String data) {
       uint16_t _gear_position         = (uint16_t) data.substring(12, 16).toInt();
       uint16_t _checksum              = (uint16_t) data.substring(MESSAGE_LENGTH - 4, MESSAGE_LENGTH).toInt();
 
-      // 0 795 1023 1023 01201102961
       Serial.println("Checking checksum components.\n");
       Serial.println(_steering_position);
       Serial.println(_brake_position);
@@ -78,9 +77,7 @@ void DataParser::parseExternalData(String data) {
         ignition_status       = (uint8_t)  data.substring(17, 18).toInt();
         kill_status           = (uint8_t)  data.substring(18, 19).toInt();
         checksum              = (uint16_t) data.substring(MESSAGE_LENGTH - 4, MESSAGE_LENGTH).toInt();
-        //Serial.println("correct sum\n");
-        //Serial.println(ignition_status);
-        //Serial.println("YOLO\n");
+
       } else {
         if (debug) {
           Serial.print(CLASS_NAME);
@@ -111,7 +108,7 @@ uint16_t DataParser::getExpectedAcceleratorPosition() {
   return accelerator_position;
 }
 
-uint8_t  DataParser::getExpectedGearPosition() {
+uint16_t  DataParser::getExpectedGearPosition() {
   return gear_position;
 }
